Move QuickSettings into Preview
