+++ feature/preview/src/main/java/com/google/jetpackcamera/feature/preview/quicksettings/ui/QuickSettingsComponents.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.feature.quicksettings.ui
+package com.google.jetpackcamera.feature.preview.quicksettings.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -45,13 +45,13 @@ import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.google.jetpackcamera.feature.quicksettings.CameraAspectRatio
-import com.google.jetpackcamera.feature.quicksettings.CameraCaptureMode
-import com.google.jetpackcamera.feature.quicksettings.CameraDynamicRange
-import com.google.jetpackcamera.feature.quicksettings.CameraFlashMode
-import com.google.jetpackcamera.feature.quicksettings.CameraLensFace
-import com.google.jetpackcamera.feature.quicksettings.QuickSettingsEnum
-import com.google.jetpackcamera.quicksettings.R
+import com.google.jetpackcamera.feature.preview.quicksettings.CameraAspectRatio
+import com.google.jetpackcamera.feature.preview.quicksettings.CameraCaptureMode
+import com.google.jetpackcamera.feature.preview.quicksettings.CameraDynamicRange
+import com.google.jetpackcamera.feature.preview.quicksettings.CameraFlashMode
+import com.google.jetpackcamera.feature.preview.quicksettings.CameraLensFace
+import com.google.jetpackcamera.feature.preview.quicksettings.QuickSettingsEnum
+import com.google.jetpackcamera.feature.preview.R
 import com.google.jetpackcamera.settings.model.AspectRatio
 import com.google.jetpackcamera.settings.model.CaptureMode
 import com.google.jetpackcamera.settings.model.DynamicRange
