+++ feature/preview/src/main/res/drawable/single_stream_capture_icon.xml
