+++ settings.gradle.kts
@@ -39,5 +39,4 @@ include(":domain:camera")
 include(":feature:settings")
 include(":data:settings")
 include(":core:common")
-include(":feature:quicksettings")
 include(":benchmark")
